+++ pyecharts/echarts/json_serializable.py
@@ -1,7 +1,10 @@
+# coding=utf8
+
+from pyecharts.interfaces import JSONSerializableMixin
 from pyecharts.utils import remove_key_with_none_value
 
 
-class JsonSerializable(object):
+class JsonSerializable(JSONSerializableMixin):
     def __init__(self):
         self._config = {}
 
@@ -14,6 +17,5 @@ class JsonSerializable(object):
     def update(self, **kwargs):
         self._config.update(**kwargs)
 
-    @property
-    def config(self):
+    def __json__(self):
         return remove_key_with_none_value(self._config)